requirements.txt
asgiref==3.8.1
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap4==2024.10
Django==5.1.3
django-crispy-forms==2.3
djangorestframework==3.15.2
dnspython==2.7.0
Flask==3.1.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
pillow==11.0.0
praw==7.8.1
prawcore==2.4.0
proto-plus==1.25.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pymongo==4.10.1
pyparsing==3.2.1
PySocks==1.7.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
sqlparse==0.5.2
trio==0.28.0
trio-websocket==0.11.1
tweepy==4.15.0
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
